+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # 0.0.3
 
 - Now support `mapStateToProps`
+- Now support setState async by default, like redux-thunk
 - Added `resetState`
 - Added `boomerang` method
 - Added `subscribe` method
 - Added `unsubscribe` method
+
 
 Fixes:
 - Added unsubscribe & remove all subscribe when umount a `withLight`